+++ cds-plugin-vector-hana/lib/index.js
@@ -2,4 +2,5 @@ module.exports = {
   VectorStore: require('./VectorStore'),
   HanaVectorStore: require('./backends/hana'),
   SqliteVectorStore: require('./backends/sqlite'),
+  RAG: require('./rag'),
 };